backend/requirements.txt
annotated-types==0.7.0
fastapi==0.115.0
uvicorn[standard]==0.32.0
neo4j==6.0.2
pydantic==2.12.3
pydantic_core==2.41.4
python-dotenv==1.0.1
pytz==2025.2
typing-inspection==0.4.2
typing_extensions==4.15.0
python-multipart==0.0.17
fastapi-cors==0.0.6
